Refactor template entity to allow reuse
